[BE] fix: 물음표 에러 해결, 반환값 변경
Former-commit-id: 3040cfc34aab74798f934b54e6ad6e0eabab308b

+++ BackEnd/SpringBootServer/src/main/java/com/ssafy/backend/domain/administration/dto/info/ClosedStoreAdministrationTopFiveInfo.java
@@ -6,6 +6,6 @@ import lombok.Builder;
 public record ClosedStoreAdministrationTopFiveInfo(
         String administrationCode,
         String administrationCodeName,
-        Double curClosedRate
+        Long curClosedRate
 ) {
 }
